Move encrypt statement to encrypt module
